+++ flutter_service_bloc/CHANGELOG.md
@@ -1,7 +1,13 @@
+## 1.0.1
+
+* revert changed onSuccess to onSucceed
+* changed onFailed to onFailure
+
 ## 1.0.0
 
 * update breaking
   dependency [service_bloc](https://github.com/aaassseee/service_bloc/blob/main/service_bloc/CHANGELOG.md)
+* changed onSuccess to onSucceed
 
 ## 0.1.0